requirements.txt
requests==2.26.0
python-telegram-bot==20.0
reportlab==3.6.1
